+++ src/Core/Application/Catalog/Products/SearchProductsRequest.cs
@@ -3,8 +3,8 @@ namespace RewardsPlus.Application.Catalog.Products;
 public class SearchProductsRequest : PaginationFilter, IRequest<PaginationResponse<ProductDto>>
 {
     public Guid? BrandId { get; set; }
-    public double? MinimumRate { get; set; }
-    public double? MaximumRate { get; set; }
+    public decimal? MinimumRate { get; set; }
+    public decimal? MaximumRate { get; set; }
 }
 
 public class SearchProductsRequestHandler : IRequestHandler<SearchProductsRequest, PaginationResponse<ProductDto>>
