+++ src/Core/Domain/CashDomain/Cash.cs
@@ -2,18 +2,18 @@ namespace RewardsPlus.Domain.CashDomain;
 
 public class Cash : AuditableEntity, IAggregateRoot
 {
-    public double Balance { get; private set; }
+    public decimal Balance { get; private set; }
     public string UserId { get; private set; }
     public string UserEmail { get; private set; }
 
-    public Cash(double balance, string userId, string userEmail)
+    public Cash(decimal balance, string userId, string userEmail)
     {
         Balance = balance;
         UserId = userId;
         UserEmail = userEmail;
     }
 
-    public Cash Update(double balance)
+    public Cash Update(decimal balance)
     {
         if (Balance != balance) Balance = balance;
         return this;
